py-agent/requirements.txt
asgiref==3.2.10
attrs==19.3.0
blindspin==2.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
crayons==0.3.1
deprecation==2.1.0
Django==3.1
docker==4.2.2
elasticsearch==7.8.0
Flask==1.1.2
gevent==20.6.2
greenlet==0.4.16
grpcio==1.31.0
grpcio-tools==1.31.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.1
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
parameterized==0.7.4
pika==1.1.0
pluggy==0.13.1
protobuf==3.12.4
py==1.9.0
pymongo==3.11.0
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==6.0.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
testcontainers==3.0.3
toml==0.10.1
tornado==6.0.4
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0

